examples/browser-use-captcha-auto/pyproject.toml
[project]
name = "steel-browser-use-captcha-solver-starter"
version = "0.1.0"
description = "Steel + Browser-Use: Automatic CAPTCHA solving"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use>=0.12.5,<0.13",
    "python-dotenv>=1.0.0",
    "steel-sdk>=0.16.0",
]
